… optuna/samplers/_tpe/multivariate_parzen_estimator.py
… tests/samplers_tests/tpe_tests/test_multivariate_parzen_estomator.py
… tests/samplers_tests/tpe_tests/test_parzen_estimator.py