Depends on #2615 and #2616.

Part of works for #2614.

## Motivation
The univariate TPE is a special case of the multivariate TPE, but the implementation of them in Optuna is overwrapped now. This PR aims to resolve the redundancy.

It seems to be difficult to support full backward compatibility including the behavior when the seed is fixed. The reason is that mus, sigmas, and weights in multivariate TPE are arranged in the order of observation, while those in current univariate TPE are arranged in the order of ascending mus. In this PR, we will adapt our logic to the multivariate TPE. We will verify through benchmarking experiments that the performance of the univariate TPE is not significantly impaired by this change.

## Description of the changes
- Unify the univariate and multivariate TPE
- Fix some tests

## TODOs
- [x] Performance benchmark on kurobako

## Benchmark Results

I took a benchmark between this PR and the current master. In summary, the changes made by this PR do not significantly impair the performance of the algorithm.

### Environments:
optuna: this PR (https://github.com/optuna/optuna/pull/2618/commits/9d25958f6eee9ab804f01ae8ca2b758e0febbfdd) and the current master (https://github.com/optuna/optuna/commit/be407fdde4533c1df3fe8ec9026e884bc9d4fb15)
python: 3.8
kurobako: 0.2.9
algorithms: multivariate-tpe-master-PRUNER, multivariate-tpe-this-PR-PRUNER, tpe-master-PRUNER, tpe-this-PR-PRUNER

Each algorithm was run 100 times with the same settings, and the mean and variance of the performance were plotted.

### Results
<details>
<summary>With NopPruner</summary>
<pre>
<code>

![hpo-bench-naval-ddecd9410805f2a70a5be7c2d5627c3b8b7af26160dbad87f5197a6a11d5f1b3](https://user-images.githubusercontent.com/38826298/119822613-d0f43c00-bf2e-11eb-80a5-4e2a4dbcffae.png)
![hpo-bench-parkinson-c54587a492ee2045a50b6b9c5db265781435c0610625a77256bad107e726d964](https://user-images.githubusercontent.com/38826298/119822619-d2bdff80-bf2e-11eb-95cd-bd4fea4bf2f1.png)
![hpo-bench-protein-71dc1cce554890df75bc2b292694102521136a26a8afd6107e56672a1f9c9e54](https://user-images.githubusercontent.com/38826298/119822620-d3569600-bf2e-11eb-9a66-9746815b9eb4.png)
![hpo-bench-slice-c4296ca79fed7b86790575cb9c5b2dab8426fd1c9b028ea209def49e73d5cb70](https://user-images.githubusercontent.com/38826298/119822621-d3ef2c80-bf2e-11eb-8295-a1d1f8cee4ad.png)
![nasbench-c-02fe653c3da0cb4ad55ff50834c60ca94a61fc7730a851507264ce322379ed19](https://user-images.githubusercontent.com/38826298/119822623-d487c300-bf2e-11eb-96a3-8d6d2ef14470.png)

</code>
</pre>
</details>

<details>
<summary>With MedianPruner</summary>
<pre>
<code>

![hpo-bench-naval-ddecd9410805f2a70a5be7c2d5627c3b8b7af26160dbad87f5197a6a11d5f1b3](https://user-images.githubusercontent.com/38826298/119822654-de112b00-bf2e-11eb-97c5-365f99a72d84.png)
![hpo-bench-parkinson-c54587a492ee2045a50b6b9c5db265781435c0610625a77256bad107e726d964](https://user-images.githubusercontent.com/38826298/119822660-df425800-bf2e-11eb-8ab2-b7641a3a072c.png)
![hpo-bench-protein-71dc1cce554890df75bc2b292694102521136a26a8afd6107e56672a1f9c9e54](https://user-images.githubusercontent.com/38826298/119822664-e0738500-bf2e-11eb-8797-03a3cdd06380.png)
![hpo-bench-slice-c4296ca79fed7b86790575cb9c5b2dab8426fd1c9b028ea209def49e73d5cb70](https://user-images.githubusercontent.com/38826298/119822670-e10c1b80-bf2e-11eb-99bd-86ad43cbe142.png)
![nasbench-c-02fe653c3da0cb4ad55ff50834c60ca94a61fc7730a851507264ce322379ed19](https://user-images.githubusercontent.com/38826298/119822671-e1a4b200-bf2e-11eb-904a-00a407d83aa8.png)

</code>
</pre>
</details>

<details>
<summary>With HyperbandPruner</summary>
<pre>
<code>

![hpo-bench-naval-ddecd9410805f2a70a5be7c2d5627c3b8b7af26160dbad87f5197a6a11d5f1b3](https://user-images.githubusercontent.com/38826298/119822703-e9645680-bf2e-11eb-9125-9d700e1a4b42.png)
![hpo-bench-parkinson-c54587a492ee2045a50b6b9c5db265781435c0610625a77256bad107e726d964](https://user-images.githubusercontent.com/38826298/119822709-eb2e1a00-bf2e-11eb-812d-1fb40cbb7c69.png)
![hpo-bench-protein-71dc1cce554890df75bc2b292694102521136a26a8afd6107e56672a1f9c9e54](https://user-images.githubusercontent.com/38826298/119822713-ebc6b080-bf2e-11eb-840c-b61b247fc747.png)
![hpo-bench-slice-c4296ca79fed7b86790575cb9c5b2dab8426fd1c9b028ea209def49e73d5cb70](https://user-images.githubusercontent.com/38826298/119822715-ec5f4700-bf2e-11eb-8a4c-eecfb5d7ac5d.png)
![nasbench-c-02fe653c3da0cb4ad55ff50834c60ca94a61fc7730a851507264ce322379ed19](https://user-images.githubusercontent.com/38826298/119822717-ecf7dd80-bf2e-11eb-8e03-73a886314670.png)

</code>
</pre>
</details>