@HideakiImamura 2021-05-13
I took a benchmark between this PR and the current master. In summary, the changes made by this PR do not significantly impair the performance of the algorithm.

### Environments:
optuna: this PR (https://github.com/optuna/optuna/pull/2618/commits/9d25958f6eee9ab804f01ae8ca2b758e0febbfdd) and the current master (https://github.com/optuna/optuna/commit/be407fdde4533c1df3fe8ec9026e884bc9d4fb15)
python: 3.8
kurobako: 0.2.9

Each algorithm was run 100 times with the same settings, and the mean and variance of the performance were plotted.

### Results
<details>
<summary>With NopPruner</summary>
<pre>
<code>

![hpo-bench-naval-ddecd9410805f2a70a5be7c2d5627c3b8b7af26160dbad87f5197a6a11d5f1b3](https://user-images.githubusercontent.com/38826298/118072559-09e7c900-b3e5-11eb-9ac8-aeb25683aa65.png)
![hpo-bench-parkinson-c54587a492ee2045a50b6b9c5db265781435c0610625a77256bad107e726d964](https://user-images.githubusercontent.com/38826298/118072563-0bb18c80-b3e5-11eb-8257-41f14976cd81.png)
![hpo-bench-protein-71dc1cce554890df75bc2b292694102521136a26a8afd6107e56672a1f9c9e54](https://user-images.githubusercontent.com/38826298/118072565-0c4a2300-b3e5-11eb-8cdb-aded0ed371f6.png)
![hpo-bench-slice-c4296ca79fed7b86790575cb9c5b2dab8426fd1c9b028ea209def49e73d5cb70](https://user-images.githubusercontent.com/38826298/118072568-0ce2b980-b3e5-11eb-9754-ceca798b3398.png)
![nasbench-a-f7734d47490690221086445e6217606ba300517d8fedeca9b94637ca6f49a6e9](https://user-images.githubusercontent.com/38826298/118072569-0ce2b980-b3e5-11eb-82d3-dbf954d7c46e.png)

</code>
</pre>
</details>

<details>
<summary>With MedianPruner</summary>
<pre>
<code>

![hpo-bench-naval-ddecd9410805f2a70a5be7c2d5627c3b8b7af26160dbad87f5197a6a11d5f1b3](https://user-images.githubusercontent.com/38826298/118072601-24ba3d80-b3e5-11eb-892e-981bfca26133.png)
![hpo-bench-parkinson-c54587a492ee2045a50b6b9c5db265781435c0610625a77256bad107e726d964](https://user-images.githubusercontent.com/38826298/118072603-26840100-b3e5-11eb-8eef-3e3d992f0f35.png)
![hpo-bench-protein-71dc1cce554890df75bc2b292694102521136a26a8afd6107e56672a1f9c9e54](https://user-images.githubusercontent.com/38826298/118072606-27b52e00-b3e5-11eb-9b42-d7fb8cdc0b5b.png)
![hpo-bench-slice-c4296ca79fed7b86790575cb9c5b2dab8426fd1c9b028ea209def49e73d5cb70](https://user-images.githubusercontent.com/38826298/118072608-284dc480-b3e5-11eb-8c81-1d1f8168c15b.png)
![nasbench-a-f7734d47490690221086445e6217606ba300517d8fedeca9b94637ca6f49a6e9](https://user-images.githubusercontent.com/38826298/118072610-284dc480-b3e5-11eb-812b-2e09147a932e.png)

</code>
</pre>
</details>

<details>
<summary>With HyperbandPruner</summary>
<pre>
<code>

![hpo-bench-naval-ddecd9410805f2a70a5be7c2d5627c3b8b7af26160dbad87f5197a6a11d5f1b3](https://user-images.githubusercontent.com/38826298/118072625-313e9600-b3e5-11eb-8225-1acae79955bd.png)
![hpo-bench-parkinson-c54587a492ee2045a50b6b9c5db265781435c0610625a77256bad107e726d964](https://user-images.githubusercontent.com/38826298/118072631-326fc300-b3e5-11eb-9779-b9e3c6518d8f.png)
![hpo-bench-protein-71dc1cce554890df75bc2b292694102521136a26a8afd6107e56672a1f9c9e54](https://user-images.githubusercontent.com/38826298/118072633-33a0f000-b3e5-11eb-904a-5a8795261b54.png)
![hpo-bench-slice-c4296ca79fed7b86790575cb9c5b2dab8426fd1c9b028ea209def49e73d5cb70](https://user-images.githubusercontent.com/38826298/118072635-34398680-b3e5-11eb-973e-fcaf145431b0.png)
![nasbench-a-f7734d47490690221086445e6217606ba300517d8fedeca9b94637ca6f49a6e9](https://user-images.githubusercontent.com/38826298/118072637-34d21d00-b3e5-11eb-89ff-5a57c1e6ab86.png)

</code>
</pre>
</details>