@hvy 2021-05-17
Could you reduce the diff now that https://github.com/optuna/optuna/pull/2615 is merged?