@HideakiImamura 2021-05-20
Thanks, @hvy. I've just merged the master branch and made this PR ready for review. PTAL.

I've noticed that I made a mistake with the benchmark. I aggregated all results of `TPESampler(multivariate=False)` and `TPESampler(multivariate=True)` in one picture. Now, I', regenerating the aggregated results.