@HideakiImamura 2021-05-20
@hvy I've included the benchmark results in the PR description. Due to my mistake, I did not have the results for MedianPruner and HyperbandPruner when multivariate=True. I will run the experiment now and add it to the description later. Sorry for the confusion.