@HideakiImamura 2021-05-20
@hvy @keisuke-umezawa I completed the benchmark experiment for univariate TPE and multivariate TPE and added those results in the PR description. Please take a look.