@HideakiImamura 2021-05-21
@hvy Thanks for your review! I resolved the conflicts.

> Are continuous parameters covered by the current benchmarks?

Yes, as in [here](https://github.com/optuna/kurobako/blob/master/kurobako_problems/src/hpobench.rs#L73), kurobako's HPO bench includes not only categorical variables but also discrete and uniform variables.