@hvy 2021-05-21
Hm, HPOBench contains discrete parameters but no continuous ones.