@HideakiImamura 2021-05-21
Ops. `Validation MSE` is not an input but an output. HPO bench includes categorical and discrete variables. Discrete variables are internally converted to real-valued variables, so the benchmark results include the continuous case.