@hvy 2021-05-21
Some benchmarks with `InMemoryStorage` to compare the speed of the the default sampler `TPESampler` without any explicit arguments. It seems like this version is somewhat slower when the number of trials increase. The difference is small when using `RDBStorage` so it might be negligible but it'll likely be noticeable for `InMemoryStorage` users so I'd like some feedback from others.

<details>
<summary>`suggest_float`</summary>

### `master`

| #params | #trials | time(sec) | time/trial(ms) | time/trial/param(ms) |
| ------- | ------- | --------- | -------------- | -------------------- |
| 1 | 1 | 0.00 | 0.51 | 0.51 |
| 1 | 10 | 0.00 | 0.18 | 0.18 |
| 1 | 100 | 0.09 | 0.94 | 0.94 |
| 1 | 1000 | 2.29 | 2.29 | 2.29 |
| 2 | 1 | 0.00 | 0.36 | 0.18 |
| 2 | 10 | 0.00 | 0.24 | 0.12 |
| 2 | 100 | 0.16 | 1.56 | 0.78 |
| 2 | 1000 | 4.45 | 4.45 | 2.22 |
| 4 | 1 | 0.00 | 0.54 | 0.14 |
| 4 | 10 | 0.00 | 0.41 | 0.10 |
| 4 | 100 | 0.32 | 3.17 | 0.79 |
| 4 | 1000 | 8.80 | 8.80 | 2.20 |
| 8 | 1 | 0.00 | 0.81 | 0.10 |
| 8 | 10 | 0.01 | 0.73 | 0.09 |
| 8 | 100 | 0.63 | 6.27 | 0.78 |
| 8 | 1000 | 17.40 | 17.40 | 2.18 |
| 16 | 1 | 0.00 | 1.36 | 0.08 |
| 16 | 10 | 0.01 | 1.38 | 0.09 |
| 16 | 100 | 1.36 | 13.56 | 0.85 |
| 16 | 1000 | 36.77 | 36.77 | 2.30 |
| 32 | 1 | 0.00 | 2.44 | 0.08 |
| 32 | 10 | 0.03 | 2.68 | 0.08 |
| 32 | 100 | 2.47 | 24.72 | 0.77 |
| 32 | 1000 | 73.25 | 73.25 | 2.29 |

### PR

| #params | #trials | time(sec) | time/trial(ms) | time/trial/param(ms) |
| ------- | ------- | --------- | -------------- | -------------------- |
| 1 | 1 | 0.00 | 0.42 | 0.42 |
| 1 | 10 | 0.00 | 0.17 | 0.17 |
| 1 | 100 | 0.11 | 1.07 | 1.07 |
| 1 | 1000 | 3.28 | 3.28 | 3.28 |
| 2 | 1 | 0.00 | 0.37 | 0.19 |
| 2 | 10 | 0.00 | 0.27 | 0.14 |
| 2 | 100 | 0.21 | 2.11 | 1.06 |
| 2 | 1000 | 6.38 | 6.38 | 3.19 |
| 4 | 1 | 0.00 | 0.95 | 0.24 |
| 4 | 10 | 0.00 | 0.48 | 0.12 |
| 4 | 100 | 0.42 | 4.25 | 1.06 |
| 4 | 1000 | 12.89 | 12.89 | 3.22 |
| 8 | 1 | 0.00 | 0.79 | 0.10 |
| 8 | 10 | 0.01 | 0.75 | 0.09 |
| 8 | 100 | 0.79 | 7.88 | 0.99 |
| 8 | 1000 | 25.74 | 25.74 | 3.22 |
| 16 | 1 | 0.00 | 1.36 | 0.09 |
| 16 | 10 | 0.01 | 1.42 | 0.09 |
| 16 | 100 | 1.62 | 16.17 | 1.01 |
| 16 | 1000 | 53.64 | 53.64 | 3.35 |
| 32 | 1 | 0.00 | 2.97 | 0.09 |
| 32 | 10 | 0.03 | 2.99 | 0.09 |
| 32 | 100 | 3.24 | 32.35 | 1.01 |
| 32 | 1000 | 107.70 | 107.70 | 3.37 |

</details>

<details>
<summary>`suggest_categorical` (3 choices)</summary>

### `master`

| #params | #trials | time(sec) | time/trial(ms) | time/trial/param(ms) |
| ------- | ------- | --------- | -------------- | -------------------- |
| 1 | 1 | 0.00 | 0.41 | 0.41 |
| 1 | 10 | 0.00 | 0.16 | 0.16 |
| 1 | 100 | 0.04 | 0.41 | 0.41 |
| 1 | 1000 | 1.64 | 1.64 | 1.64 |
| 2 | 1 | 0.00 | 0.34 | 0.17 |
| 2 | 10 | 0.00 | 0.23 | 0.11 |
| 2 | 100 | 0.07 | 0.73 | 0.37 |
| 2 | 1000 | 3.15 | 3.15 | 1.57 |
| 4 | 1 | 0.00 | 0.48 | 0.12 |
| 4 | 10 | 0.00 | 0.36 | 0.09 |
| 4 | 100 | 0.14 | 1.37 | 0.34 |
| 4 | 1000 | 6.21 | 6.21 | 1.55 |
| 8 | 1 | 0.00 | 0.71 | 0.09 |
| 8 | 10 | 0.01 | 0.67 | 0.08 |
| 8 | 100 | 0.25 | 2.52 | 0.31 |
| 8 | 1000 | 12.24 | 12.24 | 1.53 |
| 16 | 1 | 0.00 | 1.22 | 0.08 |
| 16 | 10 | 0.01 | 1.24 | 0.08 |
| 16 | 100 | 0.49 | 4.92 | 0.31 |
| 16 | 1000 | 25.26 | 25.26 | 1.58 |
| 32 | 1 | 0.00 | 2.15 | 0.07 |
| 32 | 10 | 0.02 | 2.41 | 0.08 |
| 32 | 100 | 0.96 | 9.65 | 0.30 |
| 32 | 1000 | 58.92 | 58.92 | 1.84 |

### PR

| #params | #trials | time(sec) | time/trial(ms) | time/trial/param(ms) |
| ------- | ------- | --------- | -------------- | -------------------- |
| 1 | 1 | 0.00 | 0.39 | 0.39 |
| 1 | 10 | 0.00 | 0.16 | 0.16 |
| 1 | 100 | 0.06 | 0.65 | 0.65 |
| 1 | 1000 | 2.38 | 2.38 | 2.38 |
| 2 | 1 | 0.00 | 0.47 | 0.24 |
| 2 | 10 | 0.01 | 0.52 | 0.26 |
| 2 | 100 | 0.13 | 1.26 | 0.63 |
| 2 | 1000 | 5.10 | 5.10 | 2.55 |
| 4 | 1 | 0.00 | 0.46 | 0.12 |
| 4 | 10 | 0.00 | 0.40 | 0.10 |
| 4 | 100 | 0.23 | 2.29 | 0.57 |
| 4 | 1000 | 9.14 | 9.14 | 2.28 |
| 8 | 1 | 0.00 | 0.72 | 0.09 |
| 8 | 10 | 0.01 | 0.72 | 0.09 |
| 8 | 100 | 0.46 | 4.62 | 0.58 |
| 8 | 1000 | 18.23 | 18.23 | 2.28 |
| 16 | 1 | 0.00 | 1.42 | 0.09 |
| 16 | 10 | 0.01 | 1.34 | 0.08 |
| 16 | 100 | 0.89 | 8.92 | 0.56 |
| 16 | 1000 | 37.43 | 37.43 | 2.34 |
| 32 | 1 | 0.00 | 2.30 | 0.07 |
| 32 | 10 | 0.03 | 2.60 | 0.08 |
| 32 | 100 | 1.80 | 18.03 | 0.56 |
| 32 | 1000 | 77.96 | 77.96 | 2.44 |

</details>

<details>
<summary>Code</summary>

```python
import argparse
import math
import time

import sqlalchemy

import optuna


class Profile:
    def __enter__(self):
        self.start = time.time()
        return self

    def __exit__(self, exc_type, exc_val, exc_tb):
        self.end = time.time()

    def get(self):
        return self.end - self.start


def build_objective_func(n_param):
    def objective(trial):
        return sum(
            [
                #math.sin(trial.suggest_float("param-{}".format(i), 0, 1))
                int(trial.suggest_categorical("param-{}".format(i), [1, 2, 3]))
                for i in range(n_param)
            ]
        )

    return objective


if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument("mysql_user", type=str)
    parser.add_argument("mysql_host", type=str)
    parser.add_argument("--n-params", type=int, nargs="+", default=[1, 2, 4, 8, 16, 32])
    parser.add_argument("--n-trials", type=int, nargs="+", default=[1, 10, 100, 1000])
    args = parser.parse_args()

    storage_str = "mysql+pymysql://{}@{}/".format(args.mysql_user, args.mysql_host)

    n_params = args.n_params
    n_trials = args.n_trials

    optuna.logging.set_verbosity(optuna.logging.CRITICAL)

    print(f"| #params | #trials | time(sec) | time/trial(ms) | time/trial/param(ms) |")
    print(f"| ------- | ------- | --------- | -------------- | -------------------- |")

    for n_param in n_params:
        for n_trial in n_trials:
            engine = sqlalchemy.create_engine(storage_str)
            conn = engine.connect()
            conn.execute("commit")
            database_str = "profile_storage_t{}_p{}".format(n_trial, n_param)
            try:
                conn.execute("drop database {}".format(database_str))
            except Exception:
                pass
            conn.execute("create database {}".format(database_str))
            conn.close()

            storage = optuna.storages.get_storage(storage_str + database_str)
            sampler = optuna.samplers.TPESampler()
            # study = optuna.create_study(storage=storage, sampler=sampler)
            study = optuna.create_study(sampler=sampler)

            with Profile() as prof:
                study.optimize(
                    build_objective_func(n_param),
                    n_trials=n_trial,
                    gc_after_trial=False,
                )

            print(
                f"| {n_param} | {n_trial} | {prof.get():.2f} "
                f"| {prof.get() / n_trial * 1e3:.2f} "
                f"| {prof.get() / n_trial / n_param * 1e3:.2f} "
                "|"
            )
```

</details>

_I have not profiled the code yet._